Remove deprecated single winner auction logic (#3489)
# Description
Since we made the switch to combinatorial auctions
(https://github.com/cowprotocol/services/pull/3423,
https://github.com/cowprotocol/services/pull/3385 and
https://github.com/cowprotocol/services/pull/3384) we no longer have the
need for the old, single winner max score, implementation.

This PR removes all logic related to the legacy implementation as well
as any support code used to allow switching between implementations
(legacy or combinatorial).

This PR does not deal with legacy database table removal, as these are
covered already by https://github.com/cowprotocol/services/pull/3366 and
https://github.com/cowprotocol/services/pull/3362

# Changes
<!-- List of detailed changes (how the change is accomplished) -->

- Use always combinatorial auction implementation on the autopilot,
without cutover logic.
- Removed legacy `max_score` implementation
- Removed the `Arbitrator` trait used to abstract and define common
operations of both `max_score` and `combinatorial` implementation. Now
all the operations are defined directly on the `combinatorial` mod.
- Removed competition logic that stored scores on the legacy
`settlement_score` table
- Removed the `local_node_fairness_check ` e2e test, as it was designed
for the legacy implementation and we are already covering that logic for
combinatorial auctions with other tests
- Adjusted the Uniswap V2 e2e test to stop relying on the legacy
database tables.

## How to test
I rely on the existing test passing, as no behavior should change.

## Related Issues
#3424

---------
