+++ crates/e2e/tests/e2e/limit_orders.rs
@@ -432,7 +432,6 @@ async fn two_limit_orders_multiple_winners_test(web3: Web3) {
             hex::encode(solver_a.address()), hex::encode(solver_b.address())),
             "--price-estimation-drivers=solver1|http://localhost:11088/test_solver".to_string(),
             "--max-winners-per-auction=2".to_string(),
-            "--combinatorial-auctions-cutover=1970-03-27T15:04:50.410Z".to_string()
         ],
     ).await;
