+++ crates/e2e/tests/e2e/solver_competition.rs
@@ -23,12 +23,6 @@ async fn local_node_solver_competition() {
 
 #[tokio::test]
 #[ignore]
-async fn local_node_fairness_check() {
-    run_test(fairness_check).await;
-}
-
-#[tokio::test]
-#[ignore]
 async fn local_node_wrong_solution_submission_address() {
     run_test(wrong_solution_submission_address).await;
 }
@@ -147,153 +141,6 @@ async fn solver_competition(web3: Web3) {
     // Winning candidate
     assert!(competition.solutions[1].ranking == 1);
     assert!(competition.solutions[1].is_winner);
-}
-
-async fn fairness_check(web3: Web3) {
-    let mut onchain = OnchainComponents::deploy(web3).await;
-
-    let [solver_a, solver_b] = onchain.make_solvers(to_wei(1)).await;
-    let [trader_a, trader_b] = onchain.make_accounts(to_wei(1)).await;
-    let [token_a, token_b] = onchain
-        .deploy_tokens_with_weth_uni_v2_pools(to_wei(1_000), to_wei(1_000))
-        .await;
-
-    // Fund traders
-    token_a.mint(trader_a.address(), to_wei(10)).await;
-    token_b.mint(trader_b.address(), to_wei(10)).await;
-
-    // Create more liquid routes between token_a (token_b) and weth via base_a
-    // (base_b). base_a has more liquidity than base_b, leading to the solver that
-    // knows about base_a to win
-    let [base_a, base_b] = onchain
-        .deploy_tokens_with_weth_uni_v2_pools(to_wei(10_000), to_wei(10_000))
-        .await;
-    onchain
-        .seed_uni_v2_pool((&token_a, to_wei(100_000)), (&base_a, to_wei(100_000)))
-        .await;
-    onchain
-        .seed_uni_v2_pool((&token_b, to_wei(10_000)), (&base_b, to_wei(10_000)))
-        .await;
-
-    // Approve GPv2 for trading
-    tx!(
-        trader_a.account(),
-        token_a.approve(onchain.contracts().allowance, to_wei(100))
-    );
-    tx!(
-        trader_b.account(),
-        token_b.approve(onchain.contracts().allowance, to_wei(100))
-    );
-
-    // Start system, with two solvers, one that knows about base_a and one that
-    // knows about base_b
-    colocation::start_driver(
-        onchain.contracts(),
-        vec![
-            colocation::start_baseline_solver(
-                "test_solver".into(),
-                solver_a.clone(),
-                onchain.contracts().weth.address(),
-                vec![base_a.address()],
-                1,
-                true,
-            )
-            .await,
-            colocation::start_baseline_solver(
-                "solver2".into(),
-                solver_b.clone(),
-                onchain.contracts().weth.address(),
-                vec![base_b.address()],
-                1,
-                true,
-            )
-            .await,
-        ],
-        colocation::LiquidityProvider::UniswapV2,
-        false,
-    );
-
-    let services = Services::new(&onchain).await;
-    services.start_autopilot(
-        None,
-        // Solver 1 has a fairness threshold of 0.01 ETH, which should be triggered by sub-optimally settling order_b
-        vec![
-            format!("--drivers=solver1|http://localhost:11088/test_solver|{}|10000000000000000,solver2|http://localhost:11088/solver2|{}", hex::encode(solver_a.address()), hex::encode(solver_b.address())),
-            "--price-estimation-drivers=solver1|http://localhost:11088/test_solver".to_string(),
-        ],
-    ).await;
-    services
-        .start_api(vec![
-            "--price-estimation-drivers=solver1|http://localhost:11088/test_solver".to_string(),
-        ])
-        .await;
-
-    // Place Orders
-    let order_a = OrderCreation {
-        sell_token: token_a.address(),
-        sell_amount: to_wei(10),
-        buy_token: onchain.contracts().weth.address(),
-        buy_amount: to_wei(5),
-        valid_to: model::time::now_in_epoch_seconds() + 300,
-        kind: OrderKind::Sell,
-        ..Default::default()
-    }
-    .sign(
-        EcdsaSigningScheme::Eip712,
-        &onchain.contracts().domain_separator,
-        SecretKeyRef::from(&SecretKey::from_slice(trader_a.private_key()).unwrap()),
-    );
-    let uid_a = services.create_order(&order_a).await.unwrap();
-
-    onchain.mint_block().await;
-
-    let order_b = OrderCreation {
-        sell_token: token_b.address(),
-        sell_amount: to_wei(10),
-        buy_token: onchain.contracts().weth.address(),
-        buy_amount: to_wei(5),
-        valid_to: model::time::now_in_epoch_seconds() + 300,
-        kind: OrderKind::Sell,
-        ..Default::default()
-    }
-    .sign(
-        EcdsaSigningScheme::Eip712,
-        &onchain.contracts().domain_separator,
-        SecretKeyRef::from(&SecretKey::from_slice(trader_b.private_key()).unwrap()),
-    );
-    services.create_order(&order_b).await.unwrap();
-
-    // Wait for trade
-    let indexed_trades = || async {
-        onchain.mint_block().await;
-        match services.get_trades(&uid_a).await.unwrap().first() {
-            Some(trade) => services
-                .get_solver_competition(trade.tx_hash.unwrap())
-                .await
-                .is_ok(),
-            None => false,
-        }
-    };
-    wait_for_condition(TIMEOUT, indexed_trades).await.unwrap();
-
-    // Verify that test_solver was excluded due to fairness check
-    let trades = services.get_trades(&uid_a).await.unwrap();
-    let competition = services
-        .get_solver_competition(trades[0].tx_hash.unwrap())
-        .await
-        .unwrap();
-    tracing::info!(?competition, "competition");
-
-    assert_eq!(competition.solutions.len(), 2);
-    let unfair_solution = &competition.solutions[0];
-    assert_eq!(unfair_solution.solver_address, solver_a.address());
-    assert!(unfair_solution.filtered_out);
-    assert!(!unfair_solution.is_winner);
-
-    let fair_solution = &competition.solutions[1];
-    assert_eq!(fair_solution.solver_address, solver_b.address());
-    assert!(!fair_solution.filtered_out);
-    assert!(fair_solution.is_winner);
 }
 
 async fn wrong_solution_submission_address(web3: Web3) {
@@ -514,7 +361,6 @@ async fn store_filtered_solutions(web3: Web3) {
                 "--price-estimation-drivers=test_solver|http://localhost:11088/test_solver"
                     .to_string(),
                 "--max-winners-per-auction=10".to_string(),
-                "--combinatorial-auctions-cutover=1970-03-27T15:04:50.410Z".to_string()
             ],
         )
         .await;
